Moving Testing functions to their own directory. Minor bugfix to MyMemZero. Updated sources.mk to reflect to new testing files.
