+++ Main/includeall.h
@@ -3,6 +3,11 @@
 #include "memory.h"
 #include "data.h"
 #include "project1.h"
+
 #ifdef FRDM
-	#define printf( ... ) ( 0 )
+#define printf( ... ) ( 0 )
+#endif
+
+#ifdef DEBUG
+#include "testing.h"
 #endif
