+++ Testing/testing.h
@@ -1,0 +1,11 @@
+#ifndef __TESTING__
+#define __TESTING__
+
+#include "includeall.h"
+
+#ifdef DEBUG
+#define PASS 0
+void Testing();
+#endif //DEBUG
+
+#endif // __TESTING__
