+++ sources.mk
@@ -3,12 +3,14 @@ INCLUDES = \
 	-I ./Data/   \
 	-I ./Memory/ \
 	-I ./Project1/ \
+	-I ./Testing/
 
 SRCS = \
 	./Main/main.c   \
 	./Memory/memory.c \
 	./Data/data.c \
-	./Project1/project1.c
+	./Project1/project1.c \
+	./Testing/testing.c
 
 LIBS = \
 	./Memory/memory.c \
